New module mock-beans
