+++ microprofile/testing/junit5/src/main/java/module-info.java
@@ -24,7 +24,6 @@ module io.helidon.microprofile.testing.junit5 {
     requires io.helidon.microprofile.cdi;
     requires jakarta.inject;
     requires org.junit.jupiter.api;
-    requires org.mockito;
 
     requires transitive jakarta.cdi;
     requires transitive jakarta.ws.rs;
