+++ microprofile/testing/testng/src/main/java/io/helidon/microprofile/testing/testng/HelidonTestNgListener.java
@@ -30,7 +30,6 @@ import java.lang.reflect.Modifier;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
@@ -50,15 +49,12 @@ import jakarta.enterprise.inject.se.SeContainer;
 import jakarta.enterprise.inject.se.SeContainerInitializer;
 import jakarta.enterprise.inject.spi.AfterBeanDiscovery;
 import jakarta.enterprise.inject.spi.AnnotatedType;
-import jakarta.enterprise.inject.spi.Bean;
 import jakarta.enterprise.inject.spi.BeanManager;
 import jakarta.enterprise.inject.spi.BeforeBeanDiscovery;
 import jakarta.enterprise.inject.spi.CDI;
 import jakarta.enterprise.inject.spi.Extension;
 import jakarta.enterprise.inject.spi.InjectionTarget;
 import jakarta.enterprise.inject.spi.InjectionTargetFactory;
-import jakarta.enterprise.inject.spi.ProcessAnnotatedType;
-import jakarta.enterprise.inject.spi.WithAnnotations;
 import jakarta.enterprise.inject.spi.configurator.AnnotatedTypeConfigurator;
 import jakarta.enterprise.util.AnnotationLiteral;
 import jakarta.inject.Inject;
@@ -70,8 +66,6 @@ import org.eclipse.microprofile.config.spi.ConfigBuilder;
 import org.eclipse.microprofile.config.spi.ConfigProviderResolver;
 import org.eclipse.microprofile.config.spi.ConfigSource;
 import org.glassfish.jersey.ext.cdi1x.internal.CdiComponentProvider;
-import org.mockito.MockSettings;
-import org.mockito.Mockito;
 import org.testng.IClassListener;
 import org.testng.ITestClass;
 import org.testng.ITestListener;
@@ -313,11 +307,6 @@ public class HelidonTestNgListener implements IClassListener, ITestListener {
                         + " injection into fields or constructor is not supported, as each"
                         + " test method uses a different CDI container. Field " + field
                         + " is annotated with @Inject");
-            } else if (field.getAnnotation(MockBean.class) != null) {
-                throw new RuntimeException("When a class is annotated with @HelidonTest(resetPerTest=true),"
-                        + " injection into fields or constructor is not supported, as each"
-                        + " test method uses a different CDI container. Field " + field
-                        + " is annotated with @MockBean");
             }
         }
 
@@ -328,11 +317,6 @@ public class HelidonTestNgListener implements IClassListener, ITestListener {
                         + " injection into fields or constructor is not supported, as each"
                         + " test method uses a different CDI container. Field " + field
                         + " is annotated with @Inject");
-            } else if (field.getAnnotation(MockBean.class) != null) {
-                throw new RuntimeException("When a class is annotated with @HelidonTest(resetPerTest=true),"
-                        + " injection into fields or constructor is not supported, as each"
-                        + " test method uses a different CDI container. Field " + field
-                        + " is annotated with @MockBean");
             }
         }
     }
@@ -461,7 +445,6 @@ public class HelidonTestNgListener implements IClassListener, ITestListener {
     private static class AddBeansExtension implements Extension {
         private final Class<?> testClass;
         private final List<AddBean> addBeans;
-        private final Set<Class<?>> mocks = new HashSet<>();
 
         private AddBeansExtension(Class<?> testClass, List<AddBean> addBeans) {
             this.testClass = testClass;
@@ -469,7 +452,7 @@ public class HelidonTestNgListener implements IClassListener, ITestListener {
         }
 
         @SuppressWarnings("unchecked")
-        void registerOtherBeans(@Observes AfterBeanDiscovery event, BeanManager beanManager) {
+        void registerOtherBeans(@Observes AfterBeanDiscovery event) {
             Client client = ClientBuilder.newClient();
 
             event.addBean()
@@ -488,26 +471,6 @@ public class HelidonTestNgListener implements IClassListener, ITestListener {
                             return client.target("http://localhost:7001");
                         }
                     });
-
-            // Register all mocks
-            mocks.forEach(type -> {
-                event.addBean()
-                    .addType(type)
-                    .scope(ApplicationScoped.class)
-                    .alternative(true)
-                    .createWith(inst -> {
-                        Set<Bean<?>> beans = beanManager.getBeans(MockSettings.class);
-                        if (!beans.isEmpty()) {
-                            Bean<?> bean = beans.iterator().next();
-                            MockSettings mockSettings = (MockSettings) beanManager.getReference(bean, MockSettings.class,
-                                    beanManager.createCreationalContext(null));
-                            return Mockito.mock(type, mockSettings);
-                        } else {
-                            return Mockito.mock(type);
-                        }
-                    })
-                    .priority(0);
-            });
         }
 
         void registerAddedBeans(@Observes BeforeBeanDiscovery event) {
@@ -532,20 +495,6 @@ public class HelidonTestNgListener implements IClassListener, ITestListener {
                     configurator.add(scope);
                 }
             }
-        }
-
-        void processMockBean(@Observes @WithAnnotations(MockBean.class) ProcessAnnotatedType<?> obj) throws Exception {
-            var configurator = obj.configureAnnotatedType();
-            configurator.fields().forEach(field -> {
-                MockBean mockBean = field.getAnnotated().getAnnotation(MockBean.class);
-                if (mockBean != null) {
-                    Field f = field.getAnnotated().getJavaMember();
-                    // Adds @Inject to be more user friendly
-                    field.add(Literal.INSTANCE);
-                    Class<?> fieldType = f.getType();
-                    mocks.add(fieldType);
-                }
-            });
         }
 
         private boolean hasBda(Class<?> value) {
@@ -704,15 +653,4 @@ public class HelidonTestNgListener implements IClassListener, ITestListener {
         }
     }
 
-    /**
-     * Supports inline instantiation of the {@link Inject} annotation.
-     */
-    private static final class Literal extends AnnotationLiteral<Inject> implements Inject {
-
-        private static final Literal INSTANCE = new Literal();
-
-        @Serial
-        private static final long serialVersionUID = 1L;
-
-    }
 }
