+++ microprofile/tests/testing/junit5/pom.xml
@@ -56,8 +56,8 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>org.mockito</groupId>
-            <artifactId>mockito-core</artifactId>
+            <groupId>io.helidon.microprofile.testing</groupId>
+            <artifactId>helidon-microprofile-testing-mock-beans</artifactId>
             <scope>test</scope>
         </dependency>
     </dependencies>
