+++ microprofile/tests/testing/junit5/src/test/java/io/helidon/microprofile/tests/testing/junit5/TestMockBeanField.java
@@ -27,7 +27,7 @@ import jakarta.ws.rs.client.WebTarget;
 
 import io.helidon.microprofile.testing.junit5.AddBean;
 import io.helidon.microprofile.testing.junit5.HelidonTest;
-import io.helidon.microprofile.testing.junit5.MockBean;
+import io.helidon.microprofile.testing.mockbeans.MockBean;
 
 import org.junit.jupiter.api.Test;
 import org.mockito.Answers;
