+++ microprofile/tests/testing/testng/pom.xml
@@ -55,8 +55,8 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>org.mockito</groupId>
-            <artifactId>mockito-core</artifactId>
+            <groupId>io.helidon.microprofile.testing</groupId>
+            <artifactId>helidon-microprofile-testing-mock-beans</artifactId>
             <scope>test</scope>
         </dependency>
     </dependencies>
