+++ microprofile/tests/testing/testng/src/test/java/io/helidon/microprofile/tests/testing/testng/TestMockBeanField.java
@@ -25,9 +25,9 @@ import jakarta.ws.rs.GET;
 import jakarta.ws.rs.Path;
 import jakarta.ws.rs.client.WebTarget;
 
+import io.helidon.microprofile.testing.mockbeans.MockBean;
 import io.helidon.microprofile.testing.testng.AddBean;
 import io.helidon.microprofile.testing.testng.HelidonTest;
-import io.helidon.microprofile.testing.testng.MockBean;
 
 import org.mockito.Answers;
 import org.mockito.MockSettings;
